requirements.txt
PyQt6
pyaudio
webrtcvad
vosk
pyttsx3
together==1.2.0
opencv-python
numpy
librosa
noisereduce
soundfile
scipy
python-dotenv
requests
pytest
pvporcupine
elevenlabs
tenacity
rapidfuzz>=3.0.0
pillow
